% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/FizzBuzz.R
\name{Fizz_Buzz}
\alias{Fizz_Buzz}
\title{FizzBuzz Game.}
\usage{
Fizz_Buzz(x)
}
\arguments{
\item{x}{A real number}
}
\value{
FizzBuzz result
}
\description{
Because this is a very simple function, my explanation is short.
}
\details{
The goal is to count as high  as you can, but for any number evenly divisible
by 3, substitute "Fizz" and  any number evenly divisible by 5, substitute
"Buzz", and if it is divisible  by both, substitute "Fizz Buzz".
So the sequence will look like  1, 2, Fizz, 4, Buzz, Fizz, 7, 8, Fizz
}
\examples{
Fizz_Buzz(5)
Fizz_Buzz(10)
}
