% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Flagstaff_Weather.R
\docType{data}
\name{Flagstaff_Weather}
\alias{Flagstaff_Weather}
\title{Weather information at Flagstaff’s Pulliam Airport from 1950 to 2019.}
\format{
a data frame with 25450 observations
\describe{
  \item{Date}{The date of observation.}
  \item{PRCP}{Precipitation.}
  \item{SNOW}{Snowfall.}
  \item{TMAX}{Maximum temperature.}
  \item{TMIN}{Minmum temperature.}
}
}
\source{
\url{www.ncdc.noaa.gov/cdo-web/search}
}
\usage{
Flagstaff_Weather
}
\description{
Weather information at Flagstaff’s Pulliam Airport from 1950 to 2019.
}
\keyword{datasets}
